ai-dev-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.0
openai==1.52.0
chromadb==0.5.0
tree-sitter==0.23.0
tree-sitter-python==0.23.0
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0
gitpython==3.1.43
rich==13.9.0
typer==0.12.0
httpx==0.27.0
aiofiles==24.1.0
pytest==8.3.0
pytest-asyncio==0.24.0

prometheus-client==0.20.0
requests==2.32.3
